refactor: move entropy check error handling to common
